requirements/base.txt
Django==1.8.5
django-model-utils==2.3.1
mysqlclient==1.3.6
djangorestframework==3.2.4
django-dynamic-preferences==0.5.4
django-filter==0.11.0
django-paypal==0.2.7
